move debug code to debug file
