+++ agent/configmgr/debug_server_off.go
@@ -2,13 +2,17 @@
 
 package configmgr
 
-import "log/slog"
+import (
+	"log/slog"
+
+	"github.com/netboxlabs/orb-agent/agent/configmgr/fleet"
+)
 
 // debugServer is a no-op stub when built without "-tags debug".
 type debugServer struct{}
 
-func startDebugServer(_ *slog.Logger, _ debugServerOpts) (*debugServer, error) {
-	return nil, nil
+func startFleetDebugServer(_ *slog.Logger, _ *fleet.AuthTokenManager, _ chan<- struct{}) *debugServer {
+	return nil
 }
 
 func (ds *debugServer) stop()        {}
